@@ -3,10 +3,10 @@
   "description": "Provides eslint support",
   "version": "4.0.0-alpha.2",
   "dependencies": {
-    "babel-eslint": "^10.0.1",
-    "eslint": "^5.12.0",
-    "eslint-loader": "^2.1.1",
-    "eslint-plugin-react": "^7.12.3"
+    "babel-eslint": "^10.0.3",
+    "eslint": "^6.6.0",
+    "eslint-loader": "^3.0.2",
+    "eslint-plugin-react": "^7.16.0"
   },
   "source": "src",
   "main": "src",
