@@ -14,10 +14,10 @@
     "@mrbuilder/optionsmanager": "^4.0.0-alpha.1",
     "@mrbuilder/plugin-browserslist": "^4.0.0-alpha.2",
     "@mrbuilder/utils": "^4.0.0-alpha.1",
-    "raw-loader": "^1.0.0",
+    "raw-loader": "^3.1.0",
     "val-loader": "^1.1.1",
-    "webpack": "^4.28.4",
-    "webpack-cli": "^3.2.1"
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.10"
   },
   "peerDependencies": {
     "@mrbuilder/cli": "^4.0.0"
