@@ -1,27 +1,22 @@
-const path                       = require('path');
-const lernaFilteredPackages      = require('./lernaFilteredPackages');
-const {existsSync, readFileSync} = require('fs');
-const {
-          stringify,
-          parseValue,
-          parseRe,
-      }                          = require('./parse');
+export {default as lernaFilteredPackages} from "./lernaFilteredPackages";
+export * from './parse';
+
+import path from 'path';
+import {existsSync, readFileSync} from "fs";
+import JSON5 from 'json5';
+
+import {
+    parseRe,
+} from './parse';
 //JSON5 allows for a lot more convienent syntax.
-const JSON5                      = require('json5');
-const {
-          SUBSCHEMA_PROJECT_DIR
-              = process.cwd(),
-          SUBSCHEMA_DEBUG
-      }                          = process.env;
 
-const asArray = Function.call.bind(
-    Array.prototype.slice);
+export const cwd = (...args: string[]): string => {
+    return path.join(process.cwd(), ...args);
+};
 
-const cwd = (...args) =>
-    path.resolve(SUBSCHEMA_PROJECT_DIR, ...args);
+export const project = cwd;
 
-const project   = cwd;
-const parseJSON = (filename) => {
+export const parseJSON = (filename: string): {} => {
     if (!existsSync(filename)) {
         return;
     }
@@ -29,7 +24,7 @@ const parseJSON = (filename) => {
     return JSON5.parse(file, parseRe);
 };
 
-function resolvePkgDir(name, file, ...rest) {
+export function resolvePkgDir(name: string, file?: string, ...rest: string[]): string {
 
     if (file === 'package.json') {
         return require.resolve(path.join(name, 'package.json'));
@@ -42,9 +37,9 @@ function resolvePkgDir(name, file, ...rest) {
     return path.resolve(require.resolve(path.join(name, 'package.json')), '..');
 }
 
-const PKG_CACHE = {};
+const PKG_CACHE: { [key: string]: string } = {};
 
-function pkg() {
+export function pkg(): string {
 
     const _pkg = cwd('package.json');
 
@@ -52,91 +47,46 @@ function pkg() {
         return PKG_CACHE[_pkg]
     }
 
-    debug(`using package`, _pkg);
     return (PKG_CACHE[_pkg] = require(_pkg));
 
 }
 
-
-function debug(...args) {
-    if (configOrBool(SUBSCHEMA_DEBUG)) {
-        console.log('[mrbuilder]', ...args);
+export function set(aobj: {} | boolean, key: string, value: any): {} | false {
+    if (aobj === false) {
+        return aobj;
     }
-}
-
-function warn() {
-    console.warn.apply(console, asArray(arguments));
-}
-
-function info() {
-    const value = configOrBool(SUBSCHEMA_DEBUG);
-    if (value > 1) {
-        console.warn.apply(console, asArray(arguments));
-    }
-}
-
-
-function set(obj, key, value) {
-    if (obj === false) {
-        return obj;
-    }
-
-    if (obj === true) {
-        obj = {};
-    }
-
-    obj        = obj || {};
+    let obj = aobj === true ? {} : aobj || {};
     const keys = key.split('.').filter(Boolean);
     const last = camelCased(keys.pop());
-    let cobj   = obj;
+    let cobj: { [key: string]: any } = obj;
     while (keys.length) {
         const c = camelCased(keys.shift(), false);
-        cobj    = cobj[c] || (cobj[c] = {});
+        cobj = cobj[c] || (cobj[c] = {});
     }
     cobj[last] = value;
     return obj;
 }
 
-function get(obj, key, def) {
+export function get(obj: {}, key: string | string[], def?: any): any {
     if (key == null) {
         return obj == null ? def : obj;
     }
     if (obj == null) {
         return def;
     }
-    const paths = Array.isArray(key) ? key.concat() : key.split('.');
-    let ret     = obj;
+    const paths: string[] = Array.isArray(key) ? key.concat() : key.split('.');
+    let ret: any = obj;
     while (paths.length) {
-        if (!obj) {
+        if (!ret) {
             return def;
         }
-        ret = obj[paths.shift()];
+        ret = ret[paths.shift()];
     }
     return ret == null ? def : ret;
 }
 
-function applyFuncs(f1, f2) {
-    f1 = f1 && (f1.default ? f1.default : f1);
-    f2 = f2 && (f2.default ? f2.default : f2);
-    if (!f2) {
-        return wrapFunc(f1);
-    }
-    if (!f1 && f2) {
-        return wrapFunc(f2);
-    }
-    if (!f1 && !f2) {
-        return null;
-    }
-    f1 = wrapFunc(f1);
-    f2 = wrapFunc(f2);
-    return function (opts, conf) {
-        //keep scope.
-        return f1.call(this, opts, (f2.call(this, opts, conf)));
-    }
-}
 
-
-function configOrBool(value, defaultValue) {
+export function configOrBool(value: string | boolean, defaultValue?: boolean): any {
     if (value == null) {
         return false;
     }
@@ -154,7 +104,7 @@ function configOrBool(value, defaultValue) {
     }
 }
 
-const camelCased = function (str, first) {
+const camelCased = function (str: string, first?: boolean): string {
     str = str.replace(/[.-]([a-z])/g, function (g) {
         return g[1] && g[1].toUpperCase();
     });
@@ -164,32 +114,36 @@ const camelCased = function (str, first) {
     return str;
 };
 
-const camelToHyphen = (str = '') => str.replace(/([A-Z])/g,
+export const camelToHyphen = (str = '') => str.replace(/([A-Z])/g,
     (g, a, i) => `${i === 0 ? '' : '-'}${g[0].toLowerCase()}`);
-const sliced        = Function.call.bind(Array.prototype.slice);
 
 
-function resolveMap(...args) {
-    return args.reduce(function (ret, key) {
+export type StringMap = { [key: string]: string };
+
+export function resolveMap(...args: string[]): StringMap {
+    return args.reduce(function (ret: StringMap, key: string) {
         ret[key] = resolvePkgDir(key);
         return ret;
     }, {});
 }
 
+type Entry = {
+    [key: string]: string[]
+}
 
-function parseEntry(entryNoParse) {
+export function parseEntry(entryNoParse?: string | string[] | { [key: string]: string | string[] }): Entry {
     if (!entryNoParse) {
         return;
     }
     if (!(typeof entryNoParse === 'string' || Array.isArray(entryNoParse))) {
-        return Object.keys(entryNoParse).reduce(function (ret, key) {
+        return Object.keys(entryNoParse).reduce(function (ret: Entry, key: string) {
             const val = entryNoParse[key];
-            ret[key]  = Array.isArray(val) ? val.map(v => enhancedResolve(v))
+            ret[key] = Array.isArray(val) ? val.map(v => enhancedResolve(v))
                 : [enhancedResolve(val)];
             return ret;
         }, {});
     }
-    let entry        = {};
+    let entry: Entry = {};
     const entryArray = Array.isArray(entryNoParse) ? entryNoParse
         : typeof entryNoParse === 'string' ? entryNoParse.split(/,\s*/)
             : entryNoParse;
@@ -197,25 +151,17 @@ function parseEntry(entryNoParse) {
 
     for (let i = 0, l = entryArray.length; i < l; i++) {
         const parts = entryArray[i].split('=', 2);
-        let key     = parts[0], value = parts[1];
+        let key = parts[0], value = parts[1];
         if (!value) {
             value = key;
-            key   = path.basename(key).replace(/([^.]*$)/, '');
+            key = path.basename(key).replace(/([^.]*$)/, '');
         }
-        if (entry[key]) {
-            if (Array.isArray(entry[key])) {
-                entry[key].push(enhancedResolve(value));
-            } else {
-                entry[key] = [entry[key], enhancedResolve(value)];
-            }
-        } else {
-            entry[key] = enhancedResolve(value);
-        }
+        entry[key] = [...(entry[key] || []), enhancedResolve(value)];
     }
     return entry;
 }
 
-const enhancedResolve  = (p, _require = require) => {
+export const enhancedResolve = (p: string, _require = require): string => {
     if (p.startsWith(path.sep)) {
         return p;
     }
@@ -224,7 +170,7 @@ const enhancedResolve  = (p, _require = require) => {
     }
     if (p.startsWith('~')) {
         const parts = p.substring(1).split(path.sep);
-        const pkg   = parts.shift();
+        const pkg = parts.shift();
         try {
             const pkgDir = path.resolve(_require.resolve(path.join(pkg, 'package.json')), '..');
 
@@ -238,14 +184,16 @@ const enhancedResolve  = (p, _require = require) => {
     }
     return p;
 };
-const regexOrFuncApply = (first, second) => {
+type RegExOrFn = ((str: string) => boolean) | RegExp;
+
+export const regexOrFuncApply = (first?: RegExOrFn, second?: RegExOrFn): RegExOrFn => {
     if (!first) {
         return second;
     }
     if (!second) {
         return first;
     }
-    return (test) => {
+    return (test: string) => {
         if (first instanceof RegExp) {
             if (first.test(test)) {
                 return true;
@@ -261,24 +209,24 @@ const regexOrFuncApply = (first, second) => {
         return second(test);
     }
 };
-
-module.exports = {
-    parseValue,
-    stringify,
-    get,
-    set,
-    camelCased,
-    project,
-    pkg,
-    configOrBool,
-    applyFuncs,
-    parseEntry,
-    enhancedResolve,
-    debug,
-    warn,
-    regexOrFuncApply,
-    parseJSON,
-    info,
-    cwd, sliced, resolveMap, resolvePkgDir, camelToHyphen, lernaFilteredPackages
-
-};
+//
+// module.exports = {
+//     parseValue,
+//     stringify,
+//     get,
+//     set,
+//     camelCased,
+//     project,
+//     pkg,
+//     configOrBool,
+//     applyFuncs,
+//     parseEntry,
+//     enhancedResolve,
+//     debug,
+//     warn,
+//     regexOrFuncApply,
+//     parseJSON,
+//     info,
+//     cwd, sliced, resolveMap, resolvePkgDir, camelToHyphen, lernaFilteredPackages
+//
+// };
