@@ -4,22 +4,22 @@ require('@mrbuilder/plugin-browserslist');
 
 const logger = optionsManager.logger('@mrbuilder/plugin-babel');
 
-const path            = require('path');
-const fs              = require('fs');
-const babelProcess    = require('./babel-process');
+const path = require('path');
+const fs = require('fs');
+const babelProcess = require('./babel-process');
 const {camelToHyphen} = require('@mrbuilder/utils');
-const mrb             = (key, def) => optionsManager.config(`@mrbuilder/plugin-babel.${key}`, def);
-const babelrc         = mrb('babelrc', true) ? path.resolve(process.cwd(), '.babelrc') : false;
+const mrb = (key, def) => optionsManager.config(`@mrbuilder/plugin-babel.${key}`, def);
+const babelrc = mrb('babelrc', true) ? path.resolve(process.cwd(), '.babelrc') : false;
 let conf;
 if (babelrc && fs.existsSync(babelrc)) {
     logger.info('using local .babelrc', babelrc);
     conf = JSON.parse(fs.readFileSync(babelrc, 'utf8'));
 } else {
     const defConf = mrb('babelConfig', `${__dirname}/babelrc.json`) || './babelrc.json';
-    logger.info('loading', optionsManager.enabled('@mrbuilder/plugin-babel'), defConf);
+    logger.debug('loading v7', defConf);
     conf = require(defConf);
 }
-let _plugins = mrb('plugins'), _presets = mrb('presets');
+let _plugins = mrb('plugins');
 if (_plugins) {
     if (_plugins === false) {
         conf.plugins = [];
@@ -27,6 +27,7 @@ if (_plugins) {
         conf.plugins = Array.isArray(_plugins) ? _plugins : [_plugins];
     }
 }
+let _presets = mrb('presets');
 if (_presets != null) {
     if (_presets === false) {
         conf.presets = [];
@@ -40,11 +41,11 @@ let useModules = mrb('useModules');
 if (useModules) {
     logger.info('allow exporting as ES6 modules');
     const envRe = /@babel\/(?:preset-)?(?:env|es2015)$/;
-    const idx   = conf.presets.findIndex(v => envRe.test(v));
+    const idx = conf.presets.findIndex(v => envRe.test(v));
     if (idx > -1) {
-        let newMod    = conf.presets[idx];
+        let newMod = conf.presets[idx];
         const [mod, c = {}] = Array.isArray(newMod) ? newMod : [newMod];
-        c.modules     = false;
+        c.modules = false;
         conf.presets.splice(idx, 1, [mod, c])
     } else {
         conf.presets.push(['@babel/preset-env', {modules: false}])
@@ -59,8 +60,8 @@ if (optionsManager.config('@mrbuilder/plugin-typescript.useBabel')) {
 
 const applyConfig = (type) => (op) => {
     const preset = camelToHyphen(Array.isArray(op) ? op[0] : op);
-    const short  = (new RegExp(`/@babel/${type}-(${preset})/|^(${preset})$`).exec(preset));
-    const conf   = mrb(short[1] || short[2]);
+    const short = (new RegExp(`/@babel/${type}-(${preset})/|^(${preset})$`).exec(preset));
+    const conf = mrb(short[1] || short[2]);
     if (conf != null) {
         if (conf === false) {
             return;
