@@ -1,6 +1,6 @@
-const CleanWebpackPlugin = require('clean-webpack-plugin');
-const path               = require('path');
-module.exports           = function (options = {}, webpack, om) {
+const {CleanWebpackPlugin} = require('clean-webpack-plugin');
+const path = require('path');
+module.exports = function (options = {}, webpack, om) {
     const paths = [];
 
     if (options.paths) {
@@ -16,8 +16,13 @@ module.exports           = function (options = {}, webpack, om) {
     }
 
     if (paths.length) {
-        const cleanOptions = Object.assign({root: process.cwd()}, options);
-        webpack.plugins.push(new CleanWebpackPlugin(paths, cleanOptions));
+
+        const cleanOptions = Object.assign({
+            root: process.cwd(),
+            dangerouslyAllowCleanPatternsOutsideProject: options.allowExternal
+        }, options);
+        delete cleanOptions.allowExternal;
+        webpack.plugins.push(new CleanWebpackPlugin(cleanOptions));
     } else {
         (this.warn || console.warn)('no directory found not adding clean plugin');
     }
